Allow reconfiguration of admin user credentials via 'juju set'
This adds the ability to update a users' password via keystoneclient.  This is
called for every run of ensure_initial_admin().

A basic config-changed hook is added to call ensure_initial_admin(), allowing:

    - Changing the current admin user's password via
      'juju set admin-password=foo'
    - Defining a new admin user in the system via
      'juju set admin-user=newadmin admin-password=foo'.

Note, when creating a new admin user,  the previous admin user is not deleted from the system.
The new admin user can manage existing users thru the dashboard or keystoneclient.